requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
importlib-metadata==6.1.0
Markdown==3.4.3
Pillow==9.5.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
zipp==3.15.0
